# src/main/java/com/bankrest/dto/UpdateUserRequest.java
package com.bankrest.dto;

import com.bankrest.entity.Role;
import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UpdateUserRequest {
    @NotNull(message = "Роль обязательна")
    private Role role;
}