# src/main/java/com/bankrest/dto/UserResponse.java
package com.bankrest.dto;

import com.bankrest.entity.Role;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

@Data
@AllArgsConstructor
@Builder
public class UserResponse {
    private Long id;
    private String username;
    private Role role;
}